remove copied version of mask analysis
