+++ lib/Analysis/MaskAnalysis.cpp
@@ -17,25 +17,25 @@ namespace mlir {
 namespace triton {
 
 LogicalResult MaskState::parse(Value operand, const Location loc,
-                               ConversionPatternRewriter &rewriter) {
+                               OpBuilder &builder) {
   if (auto op = operand.getDefiningOp<arith::ConstantOp>()) {
-    return this->parseConstant(op, loc, rewriter);
+    return this->parseConstant(op, loc, builder);
   } else if (operand.getType().isa<IntegerType>()) {
-    return this->parseIntScalar(operand, loc, rewriter);
+    return this->parseIntScalar(operand, loc, builder);
   } else if (auto op = operand.getDefiningOp<arith::AddIOp>()) {
-    return this->parseAdd(op, loc, rewriter);
+    return this->parseAdd(op, loc, builder);
   } else if (auto op = operand.getDefiningOp<arith::AndIOp>()) {
-    return this->parseAnd(op, loc, rewriter);
+    return this->parseAnd(op, loc, builder);
   } else if (auto op = operand.getDefiningOp<arith::CmpIOp>()) {
-    return this->parseCmp(op, loc, rewriter);
+    return this->parseCmp(op, loc, builder);
   } else if (auto op = operand.getDefiningOp<triton::MakeRangeOp>()) {
-    return this->parseMakeRange(op, loc, rewriter);
+    return this->parseMakeRange(op, loc, builder);
   } else if (auto op = operand.getDefiningOp<triton::BroadcastOp>()) {
-    return this->parseBroadcast(op, loc, rewriter);
+    return this->parseBroadcast(op, loc, builder);
   } else if (auto op = operand.getDefiningOp<triton::SplatOp>()) {
-    return this->parseSplat(op, loc, rewriter);
+    return this->parseSplat(op, loc, builder);
   } else if (auto op = operand.getDefiningOp<triton::ExpandDimsOp>()) {
-    return this->parseExpandDims(op, loc, rewriter);
+    return this->parseExpandDims(op, loc, builder);
   } else {
     return failure();
   }
@@ -43,28 +43,28 @@ LogicalResult MaskState::parse(Value operand, const Location loc,
 
 tensor::ExtractSliceOp
 MaskState::getExtractSlice(Value source, const Location loc,
-                           ConversionPatternRewriter &rewriter) const {
+                           OpBuilder &builder) const {
   auto sourceType = source.getType().cast<RankedTensorType>();
-  SmallVector<OpFoldResult> offsets(getRank(), rewriter.getIndexAttr(0));
-  SmallVector<OpFoldResult> strides(getRank(), rewriter.getIndexAttr(1));
+  SmallVector<OpFoldResult> offsets(getRank(), builder.getIndexAttr(0));
+  SmallVector<OpFoldResult> strides(getRank(), builder.getIndexAttr(1));
 
   auto dstType = tensor::ExtractSliceOp::inferResultType(sourceType, offsets,
                                                          dims, strides);
 
-  return rewriter.create<tensor::ExtractSliceOp>(loc, dstType, source, offsets,
+  return builder.create<tensor::ExtractSliceOp>(loc, dstType, source, offsets,
                                                  dims, strides);
 }
 
 memref::SubViewOp
 MaskState::getSubview(Value source, const Location loc,
-                      ConversionPatternRewriter &rewriter) const {
+                      OpBuilder &builder) const {
   auto sourceType = source.getType().cast<MemRefType>();
-  SmallVector<OpFoldResult> offsets(getRank(), rewriter.getIndexAttr(0));
-  SmallVector<OpFoldResult> strides(getRank(), rewriter.getIndexAttr(1));
+  SmallVector<OpFoldResult> offsets(getRank(), builder.getIndexAttr(0));
+  SmallVector<OpFoldResult> strides(getRank(), builder.getIndexAttr(1));
   auto dstType =
       memref::SubViewOp::inferResultType(sourceType, offsets, dims, strides);
 
-  return rewriter.create<memref::SubViewOp>(loc, dstType.cast<MemRefType>(),
+  return builder.create<memref::SubViewOp>(loc, dstType.cast<MemRefType>(),
                                             source, offsets, dims, strides);
 }
 
@@ -136,20 +136,20 @@ static memref::SubViewOp createSubview(Value src, Location loc, OpBuilder &b,
 // + rowView1 = rowView2 = row = rowFull
 std::pair<memref::SubViewOp, memref::SubViewOp>
 MaskState::getSideBySideSubviews(Value block1, Value block2, const Location loc,
-                                 ConversionPatternRewriter &rewriter) const {
+                                 OpBuilder &builder) const {
   OpFoldResult subviewRowFull = dims[0];
   OpFoldResult subviewColFull = dims[1];
   OpFoldResult col1 =
-      rewriter.create<memref::DimOp>(loc, block1, 1).getResult();
-  OpFoldResult subviewCol1 = minOFRs(col1, subviewColFull, loc, rewriter);
+      builder.create<memref::DimOp>(loc, block1, 1).getResult();
+  OpFoldResult subviewCol1 = minOFRs(col1, subviewColFull, loc, builder);
   OpFoldResult subviewCol2 =
-      subOFRs(subviewColFull, subviewCol1, loc, rewriter);
+      subOFRs(subviewColFull, subviewCol1, loc, builder);
 
-  SmallVector<OpFoldResult> offsets(getRank(), rewriter.getIndexAttr(0));
-  SmallVector<OpFoldResult> strides(getRank(), rewriter.getIndexAttr(1));
-  auto sv1 = createSubview(block1, loc, rewriter, offsets,
+  SmallVector<OpFoldResult> offsets(getRank(), builder.getIndexAttr(0));
+  SmallVector<OpFoldResult> strides(getRank(), builder.getIndexAttr(1));
+  auto sv1 = createSubview(block1, loc, builder, offsets,
                            {subviewRowFull, subviewCol1}, strides);
-  auto sv2 = createSubview(block2, loc, rewriter, offsets,
+  auto sv2 = createSubview(block2, loc, builder, offsets,
                            {subviewRowFull, subviewCol2}, strides);
 
   return {sv1, sv2};
@@ -157,36 +157,36 @@ MaskState::getSideBySideSubviews(Value block1, Value block2, const Location loc,
 
 std::pair<memref::SubViewOp, memref::SubViewOp>
 MaskState::getStackedSubviews(Value block1, Value block2, const Location loc,
-                              ConversionPatternRewriter &rewriter) const {
+                              OpBuilder &builder) const {
   OpFoldResult subviewRowFull = dims[0];
   OpFoldResult subviewColFull = dims[1];
   OpFoldResult row1 =
-      rewriter.create<memref::DimOp>(loc, block1, 0).getResult();
-  OpFoldResult subviewRow1 = minOFRs(row1, subviewRowFull, loc, rewriter);
+      builder.create<memref::DimOp>(loc, block1, 0).getResult();
+  OpFoldResult subviewRow1 = minOFRs(row1, subviewRowFull, loc, builder);
   OpFoldResult subviewRow2 =
-      subOFRs(subviewRowFull, subviewRow1, loc, rewriter);
+      subOFRs(subviewRowFull, subviewRow1, loc, builder);
 
-  SmallVector<OpFoldResult> offsets(getRank(), rewriter.getIndexAttr(0));
-  SmallVector<OpFoldResult> strides(getRank(), rewriter.getIndexAttr(1));
-  auto sv1 = createSubview(block1, loc, rewriter, offsets,
+  SmallVector<OpFoldResult> offsets(getRank(), builder.getIndexAttr(0));
+  SmallVector<OpFoldResult> strides(getRank(), builder.getIndexAttr(1));
+  auto sv1 = createSubview(block1, loc, builder, offsets,
                            {subviewRow1, subviewColFull}, strides);
-  auto sv2 = createSubview(block2, loc, rewriter, offsets,
+  auto sv2 = createSubview(block2, loc, builder, offsets,
                            {subviewRow2, subviewColFull}, strides);
   return {sv1, sv2};
 }
 
 LogicalResult MaskState::addStateScalar(const MaskState &state,
                                         const OpFoldResult scalar, Location loc,
-                                        ConversionPatternRewriter &rewriter) {
-  start = addOFRs(state.start, scalar, loc, rewriter);
-  end = addOFRs(state.end, scalar, loc, rewriter);
+                                        OpBuilder &builder) {
+  start = addOFRs(state.start, scalar, loc, builder);
+  end = addOFRs(state.end, scalar, loc, builder);
   dims = state.dims;
   return success();
 }
 
 LogicalResult MaskState::addStates(const MaskState &lhsState,
                                    const MaskState &rhsState, Location loc,
-                                   ConversionPatternRewriter &rewriter) {
+                                   OpBuilder &builder) {
   if (lhsState.scalar && rhsState.scalar) {
     InFlightDiagnostic diag =
         emitError(loc) << "Unexpected case where both lhs and rhs are scalars";
@@ -201,14 +201,14 @@ LogicalResult MaskState::addStates(const MaskState &lhsState,
   }
 
   if (lhsState.scalar)
-    return addStateScalar(rhsState, lhsState.scalar, loc, rewriter);
+    return addStateScalar(rhsState, lhsState.scalar, loc, builder);
   else
-    return addStateScalar(lhsState, rhsState.scalar, loc, rewriter);
+    return addStateScalar(lhsState, rhsState.scalar, loc, builder);
 }
 
 LogicalResult MaskState::minStates(const MaskState &lhsState,
                                    const MaskState &rhsState, Location loc,
-                                   ConversionPatternRewriter &rewriter) {
+                                   OpBuilder &builder) {
   if (lhsState.getRank() != rhsState.getRank()) {
     InFlightDiagnostic diag =
         emitError(loc)
@@ -219,14 +219,13 @@ LogicalResult MaskState::minStates(const MaskState &lhsState,
   for (uint32_t i = 0; i < lhsState.getRank(); i++) {
     auto lhsDim = lhsState.dims[i];
     auto rhsDim = rhsState.dims[i];
-    dims.push_back(minOFRs(lhsDim, rhsDim, loc, rewriter));
+    dims.push_back(minOFRs(lhsDim, rhsDim, loc, builder));
   }
   return success();
 }
 
 LogicalResult MaskState::parseConstant(arith::ConstantOp constOp,
-                                       const Location loc,
-                                       ConversionPatternRewriter &rewriter) {
+                                       const Location loc, OpBuilder &builder) {
   assert(this->isEmpty());
 
   if (isa<DenseElementsAttr>(constOp.getValue())) {
@@ -236,61 +235,61 @@ LogicalResult MaskState::parseConstant(arith::ConstantOp constOp,
            "All elements must share a single integer constant value");
     auto values = attr.getValues<IntegerAttr>();
     auto value = values[0].getValue();
-    auto constAttr = rewriter.getIndexAttr(value.getSExtValue());
-    auto op = arith::ConstantOp::materialize(rewriter, constAttr,
-                                             rewriter.getIndexType(), loc);
+    auto constAttr = builder.getIndexAttr(value.getSExtValue());
+    auto op = arith::ConstantOp::materialize(builder, constAttr,
+                                             builder.getIndexType(), loc);
     this->scalar = op.getValue();
   } else {
     auto value = constOp.getValue().cast<IntegerAttr>().getInt();
-    this->scalar = rewriter.getIndexAttr(value);
+    this->scalar = builder.getIndexAttr(value);
   }
 
   return success();
 }
 
 LogicalResult MaskState::parseIntScalar(Value scalar, const Location loc,
-                                        ConversionPatternRewriter &rewriter) {
+                                        OpBuilder &builder) {
   assert(this->isEmpty());
   auto castOp =
-      rewriter.create<arith::IndexCastOp>(loc, rewriter.getIndexType(), scalar);
+      builder.create<arith::IndexCastOp>(loc, builder.getIndexType(), scalar);
   this->scalar = castOp.getResult();
   return success();
 }
 
 LogicalResult MaskState::parseAdd(arith::AddIOp addOp, const Location loc,
-                                  ConversionPatternRewriter &rewriter) {
+                                  OpBuilder &builder) {
   assert(this->isEmpty());
 
   MaskState lhsState;
-  if (failed(lhsState.parse(addOp.getLhs(), loc, rewriter)))
+  if (failed(lhsState.parse(addOp.getLhs(), loc, builder)))
     return failure();
 
   MaskState rhsState;
-  if (failed(rhsState.parse(addOp.getRhs(), loc, rewriter)))
+  if (failed(rhsState.parse(addOp.getRhs(), loc, builder)))
     return failure();
 
-  return this->addStates(lhsState, rhsState, loc, rewriter);
+  return this->addStates(lhsState, rhsState, loc, builder);
 }
 
 LogicalResult MaskState::parseAnd(arith::AndIOp andOp, const Location loc,
-                                  ConversionPatternRewriter &rewriter) {
+                                  OpBuilder &builder) {
   assert(this->isEmpty());
 
   MaskState lhsState;
-  if (failed(lhsState.parse(andOp.getLhs(), loc, rewriter)) ||
+  if (failed(lhsState.parse(andOp.getLhs(), loc, builder)) ||
       !lhsState.isMask())
     return failure();
 
   MaskState rhsState;
-  if (failed(rhsState.parse(andOp.getRhs(), loc, rewriter)) ||
+  if (failed(rhsState.parse(andOp.getRhs(), loc, builder)) ||
       !rhsState.isMask())
     return failure();
 
-  return this->minStates(lhsState, rhsState, loc, rewriter);
+  return this->minStates(lhsState, rhsState, loc, builder);
 }
 
 LogicalResult MaskState::parseCmp(arith::CmpIOp cmpOp, const Location loc,
-                                  ConversionPatternRewriter &rewriter) {
+                                  OpBuilder &builder) {
   assert(this->isEmpty());
 
   if (cmpOp.getPredicate() != arith::CmpIPredicate::slt) {
@@ -299,11 +298,11 @@ LogicalResult MaskState::parseCmp(arith::CmpIOp cmpOp, const Location loc,
   }
 
   MaskState lhsState;
-  if (failed(lhsState.parse(cmpOp.getLhs(), loc, rewriter)))
+  if (failed(lhsState.parse(cmpOp.getLhs(), loc, builder)))
     return failure();
 
   MaskState rhsState;
-  if (failed(rhsState.parse(cmpOp.getRhs(), loc, rewriter)))
+  if (failed(rhsState.parse(cmpOp.getRhs(), loc, builder)))
     return failure();
 
   assert((!lhsState.scalar && rhsState.scalar) && "Unsupported cmpi scenario");
@@ -324,8 +323,8 @@ LogicalResult MaskState::parseCmp(arith::CmpIOp cmpOp, const Location loc,
   assert(cmpDim != -1 &&
          "Unexpected case where no dimension has size larger than 1");
 
-  auto newEnd = minOFRs(lhsState.end, rhsState.scalar, loc, rewriter);
-  auto newDim = subOFRs(newEnd, lhsState.start, loc, rewriter);
+  auto newEnd = minOFRs(lhsState.end, rhsState.scalar, loc, builder);
+  auto newDim = subOFRs(newEnd, lhsState.start, loc, builder);
 
   for (int32_t i = 0; i < lhsState.getRank(); i++) {
     if (i == cmpDim)
@@ -339,7 +338,7 @@ LogicalResult MaskState::parseCmp(arith::CmpIOp cmpOp, const Location loc,
 
 LogicalResult MaskState::parseMakeRange(triton::MakeRangeOp rangeOp,
                                         const Location loc,
-                                        ConversionPatternRewriter &rewriter) {
+                                        OpBuilder &builder) {
   assert(this->isEmpty());
 
   auto shape = rangeOp.getType().cast<ShapedType>().getShape();
@@ -355,16 +354,16 @@ LogicalResult MaskState::parseMakeRange(triton::MakeRangeOp rangeOp,
     return failure();
   }
 
-  this->start = rewriter.getIndexAttr(start);
-  this->end = rewriter.getIndexAttr(end);
-  this->dims.push_back(rewriter.getIndexAttr(shape[0]));
+  this->start = builder.getIndexAttr(start);
+  this->end = builder.getIndexAttr(end);
+  this->dims.push_back(builder.getIndexAttr(shape[0]));
 
   return success();
 }
 
 LogicalResult MaskState::parseBroadcast(triton::BroadcastOp broadcastOp,
                                         const Location loc,
-                                        ConversionPatternRewriter &rewriter) {
+                                        OpBuilder &builder) {
   assert(this->isEmpty());
 
   auto src = broadcastOp.getSrc();
@@ -377,14 +376,14 @@ LogicalResult MaskState::parseBroadcast(triton::BroadcastOp broadcastOp,
   assert(srcShape.size() == dstShape.size() &&
          "rank of source and destination should match");
 
-  if (failed(parse(src, loc, rewriter)))
+  if (failed(parse(src, loc, builder)))
     return failure();
 
   for (size_t i = 0; i < srcShape.size(); i++) {
     if (srcShape[i] == dstShape[i])
       continue;
     else if (srcShape[i] < dstShape[i])
-      this->dims[i] = rewriter.getIndexAttr(dstShape[i]);
+      this->dims[i] = builder.getIndexAttr(dstShape[i]);
     else
       llvm_unreachable("unexpected dimensions used in broadcast");
   }
@@ -393,7 +392,7 @@ LogicalResult MaskState::parseBroadcast(triton::BroadcastOp broadcastOp,
 }
 
 LogicalResult MaskState::parseSplat(triton::SplatOp splatOp, const Location loc,
-                                    ConversionPatternRewriter &rewriter) {
+                                    OpBuilder &builder) {
   assert(this->isEmpty());
 
   auto src = splatOp.getSrc();
@@ -407,21 +406,21 @@ LogicalResult MaskState::parseSplat(triton::SplatOp splatOp, const Location loc,
     return failure();
   }
 
-  if (failed(this->parse(src, loc, rewriter)))
+  if (failed(this->parse(src, loc, builder)))
     return failure();
 
   for (auto s : dstShape)
-    this->dims.push_back(rewriter.getIndexAttr(s));
+    this->dims.push_back(builder.getIndexAttr(s));
 
   return success();
 }
 
 LogicalResult MaskState::parseExpandDims(triton::ExpandDimsOp expandDimsOp,
                                          const Location loc,
-                                         ConversionPatternRewriter &rewriter) {
+                                         OpBuilder &builder) {
   assert(this->isEmpty());
 
-  if (failed(this->parse(expandDimsOp.getSrc(), loc, rewriter)))
+  if (failed(this->parse(expandDimsOp.getSrc(), loc, builder)))
     return failure();
 
   auto dstShape =
@@ -429,7 +428,7 @@ LogicalResult MaskState::parseExpandDims(triton::ExpandDimsOp expandDimsOp,
   auto axis = expandDimsOp.getAxis();
   assert(dstShape[axis] == 1 &&
          "expect changed dimension to be 1 in expand_dims");
-  this->dims.insert(this->dims.begin() + axis, rewriter.getIndexAttr(1));
+  this->dims.insert(this->dims.begin() + axis, builder.getIndexAttr(1));
 
   return success();
 }
