+++ lib/AnalysisStructured/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_mlir_library(TritonSharedAnalysisStructured
-  MaskAnalysis.cpp
   PtrAnalysis.cpp
 
   DEPENDS
